refactor(agent): move the agent manager and A2A agent service into agent/application
AgentManager implements agentdomain.AgentManager and A2AAgentService implements
agentapp.A2AAgentService - both belong in the context that declares their port.
agents.go becomes a2a_agent_service.go, matching what it holds.

The struct A2AAgentService collides with the interface of the same name already
declared in a2a.go, so it is renamed A2AAgentServiceImpl, following the repo's
existing FileServiceImpl and A2ATaskTrackerImpl. NewA2AAgentService keeps its
name; the concrete type assertion in monitorExternalAgents follows the rename.
The agentapp self-import is dropped and its references unqualified.
