+++ internal/agent/application/agent_manager_test.go
@@ -1,4 +1,4 @@
-package services
+package application
 
 import (
 	"context"
